split show while creating e2e tests
