+++ tests/e2e/replication/CMakeLists.txt
@@ -13,6 +13,9 @@ copy_e2e_python_files(replication common.py)
 copy_e2e_python_files(replication conftest.py)
 copy_e2e_python_files(replication show.py)
 copy_e2e_python_files(replication show_while_creating_invalid_state.py)
+copy_e2e_python_files(replication show_while_creating_invalid_state_part_2.py)
+copy_e2e_python_files(replication show_while_creating_invalid_state_part_3.py)
+copy_e2e_python_files(replication show_while_creating_invalid_state_part_4.py)
 copy_e2e_python_files(replication edge_delete.py)
 copy_e2e_python_files(replication replicate_enum.py)
 copy_e2e_python_files(replication replicate_periodic_commit.py)
